Folder structure update: add constants folder and move constants.ts to constants folder
